add domain page error panel
